+++ AutoPark/Models/Other/Rent.cs
@@ -10,12 +10,12 @@ namespace AutoPark.Models.Other
     {
         public int VehicleId { get; set; }
         public DateTime RentDate { get; set; }
-        public double Cost { get; set; }
+        public decimal Cost { get; set; }
         public Rent()
         {
             RentDate = DateTime.Now;
         }
-        public Rent(int vehicleId, DateTime time, double cost)
+        public Rent(int vehicleId, DateTime time, decimal cost)
         {
             VehicleId = vehicleId;
             RentDate = time;
